requirements.txt
telethon==1.28.5
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2 